requirements.txt
pytest==6.2.4
SQLAlchemy==1.2.12
pyyaml>=5.3